requirements.txt
mido
python-rtmidi
pyaudio
python-dotenv